project-swach-workers/backend/requirements.txt
# API Framework
fastapi
uvicorn

# Database & ODM
beanie
motor
pydantic-settings

# Utilities
python-multipart
cloudinary
shapely
python-dotenv

# AI & Vision
google-genai
